std: Move swap and replace to prim
