+++ src/libprim/lib.rs
@@ -35,3 +35,5 @@ pub mod intrinsics;
 #[cfg(not(test))] pub mod kinds;
 pub mod mem;
 pub mod ptr;
+pub mod util;
+pub mod tuple;
