+++ src/libstd/util.rs
@@ -10,48 +10,13 @@
 
 //! Miscellaneous helpers for common patterns
 
-use cast;
-use ptr;
 use prelude::*;
-use unstable::intrinsics;
+
+pub use prim::util::{swap, replace};
 
 /// The identity function.
 #[inline]
 pub fn id<T>(x: T) -> T { x }
-
-/**
- * Swap the values at two mutable locations of the same type, without
- * deinitialising or copying either one.
- */
-#[inline]
-pub fn swap<T>(x: &mut T, y: &mut T) {
-    unsafe {
-        // Give ourselves some scratch space to work with
-        let mut tmp: T = intrinsics::uninit();
-        let t: *mut T = &mut tmp;
-
-        // Perform the swap, `&mut` pointers never alias
-        let x_raw: *mut T = x;
-        let y_raw: *mut T = y;
-        ptr::copy_nonoverlapping_memory(t, &*x_raw, 1);
-        ptr::copy_nonoverlapping_memory(x, &*y_raw, 1);
-        ptr::copy_nonoverlapping_memory(y, &*t, 1);
-
-        // y and t now point to the same thing, but we need to completely forget `tmp`
-        // because it's no longer relevant.
-        cast::forget(tmp);
-    }
-}
-
-/**
- * Replace the value at a mutable location with a new one, returning the old
- * value, without deinitialising or copying either one.
- */
-#[inline]
-pub fn replace<T>(dest: &mut T, mut src: T) -> T {
-    swap(dest, &mut src);
-    src
-}
 
 /// A non-copyable dummy type.
 #[deriving(Eq, TotalEq, Ord, TotalOrd)]
